Jetson Compatibility
testcamera.py is a python script to be run on the Jetson to ensure that the camera is recognized and is working. If it works correctly, a window with a live stream from the camera should pop up and run.
JetCode.py is an updated and improved version of Code.py (as of 2-11-2018) that works on the Jetson with a live stream of the camera. The original Code.py wasn't able to handle images that didn't have boxes in them at all. JetCode.py accounts for the situation where there are no boxes in the image. JetCode.py also takes in the camera stream. Most indents in the code are reformatted to tabs.